+++ generators/app/tasks/appSetup/featuresFiles/onBoardingFeatureFiles.js
@@ -1,13 +1,8 @@
-const { copyFile, copyTemplateFile } = require('../utils');
-const { AUTH_SERVICE, ONBOARDING_PATH, ONBOARDING_SERVICE } = require('../files');
+const { copyFile } = require('../utils');
+const { ONBOARDING_PATH } = require('../files');
 
-const FILES = [ONBOARDING_PATH, ONBOARDING_SERVICE];
-
-const TEMPLATE_FILES = [AUTH_SERVICE];
+const FILES = [ONBOARDING_PATH];
 
 module.exports = function onBoardingFeatureFiles() {
-  if (!this.features.loginandsignup) {
-    TEMPLATE_FILES.forEach(copyTemplateFile.bind(this));
-  }
   FILES.forEach(copyFile.bind(this));
 };