+++ generators/app/tasks/nextSteps.js
@@ -1,7 +1,5 @@
 require('colors');
 
-const { PUSH_NOTIFICATIONS_SETUP_LINK } = require('../constants');
-
 module.exports = function nextSteps() {
   console.log(`\n       ${'NEXT STEPS!'.bold.underline.white}       \n`);
   // Introduction message
@@ -16,19 +14,11 @@ module.exports = function nextSteps() {
   }
 
   // Firebase
-  if (this.features.crashlytics) {
+  if (this.features.firebasecrashlytics) {
     console.log(
       '  (*) Remember to ask your team lead for the google services file, this file must be replaced for the corresponding application of https://console.firebase.google.com'
         .cyan
     );
-  }
-  if (this.features.pushnotifications) {
-    // Push Notifications SetUp
-    console.log(
-      '● The push notification config has been added to the project but you still need to manually link the ios library.'
-        .cyan
-    );
-    console.log(`  Further instructions can be found here: ${PUSH_NOTIFICATIONS_SETUP_LINK.underline}`.cyan);
   }
 
   if (this.features.loginandsignup) {