+++ generators/app/templates/src/constants/fonts.ts
@@ -2,7 +2,7 @@
 // You can add here your custom fonts.
 // If you want to read more about naming and linking custom fonts,
 // this is the original post: https://medium.com/@mehran.khan/ultimate-guide-to-use-custom-fonts-in-react-native-77fcdf859cf4
-export const MONTSERRAT = 'Montserrat';
+export const NUNITO = 'Nunito';
 // WEIGHTS
 export const REGULAR = 'Regular';
 export const SEMIBOLD = 'SemiBold';